Add files via upload
I added a new ID column in the 'Simulated Transactions Data' sheet but I kept the old ID column as another column. The old ID column is used to assess the distribution of the IDs. The new column was added to randomize the order of the IDs so that similar IDs do not necessarily have very similar customer values.  I also deleted unnecessary rows at the end of one of the tables. This made things render faster.